create new components and refactor old ones
