+++ src/app/app-routing.module.ts
@@ -11,8 +11,9 @@ import { TpaViewComponent } from "./pages/tpa-management/tpa-view/tpa-view.compo
 import { TpaEditComponent } from "./pages/tpa-management/tpa-edit/tpa-edit.component";
 import {SectionsComponent} from "./pages/tpa-management/tpa-edit/sections/sections.component";
 import {AllComponent} from "./pages/tpa-management/tpa-edit/all/all.component";
+import {CloneComponent} from "./pages/gh-simulator/clone/clone.component";
+import {RepositoryComponent} from "./pages/gh-simulator/clone/repository/repository.component";
 import {GhSimulatorComponent} from "./pages/gh-simulator/gh-simulator.component";
-import {RepositoryComponent} from "./pages/gh-simulator/repository/repository.component";
 const routes: Routes = [
   {
     path: '',
@@ -61,6 +62,10 @@ const routes: Routes = [
   {
     path: 'gh-simulator',
     component: GhSimulatorComponent
+  },
+  {
+    path: 'gh-simulator/clone',
+    component: CloneComponent
   },
   {
     path: 'gh-simulator/repository/:owner/:repoName',
