+++ src/app/app.module.ts
@@ -21,8 +21,11 @@ import { TpaEditComponent } from './pages/tpa-management/tpa-edit/tpa-edit.compo
 import { TpaViewComponent } from './pages/tpa-management/tpa-view/tpa-view.component';
 import { AllComponent } from './pages/tpa-management/tpa-edit/all/all.component';
 import { SectionsComponent } from './pages/tpa-management/tpa-edit/sections/sections.component';
-import { GhSimulatorComponent } from './pages/gh-simulator/gh-simulator.component';
-import { RepositoryComponent } from './pages/gh-simulator/repository/repository.component';
+import { CloneComponent } from './pages/gh-simulator/clone/clone.component';
+import { RepositoryComponent } from './pages/gh-simulator/clone/repository/repository.component';
+import {GhSimulatorComponent} from "./pages/gh-simulator/gh-simulator.component";
+import { LocalComponent } from './pages/gh-simulator/local/local.component';
+import { ActionsComponent } from './pages/gh-simulator/local/actions/actions.component';
 
 @NgModule({
   declarations: [
@@ -40,8 +43,11 @@ import { RepositoryComponent } from './pages/gh-simulator/repository/repository.
     TpaViewComponent,
     AllComponent,
     SectionsComponent,
-    GhSimulatorComponent,
+    CloneComponent,
     RepositoryComponent,
+    GhSimulatorComponent,
+    LocalComponent,
+    ActionsComponent
   ],
   imports: [
     BrowserModule,
