+++ src/app/pages/gh-simulator/clone/clone.component.css
@@ -1,0 +1,54 @@
+.token-section div {
+  margin: 20px;
+  padding: 20px;
+  border: 1px solid #ddd;
+  border-radius: 5px;
+  background-color: #f9f9f9;
+}
+
+.token-section p {
+  font-size: 18px;
+  color: #333;
+}
+
+.token-section button {
+  margin-top: 10px;
+  padding: 10px 20px;
+  border: none;
+  border-radius: 5px;
+  background-color: #007bff;
+  color: #fff;
+  cursor: pointer;
+}
+
+.token-section button:hover {
+  background-color: #0056b3;
+}
+#volver {
+  margin-left: auto;
+  float: right;
+  clear: both;
+  margin-right: 40px;
+  padding: 10px 20px;
+  font-size: 20px;
+  width: 100px;
+  height: 50px;
+  margin-top: 20px;
+
+}
+.token-section input {
+  margin-top: 10px;
+  padding: 10px;
+  border: 1px solid #ddd;
+  border-radius: 5px;
+  width: 100%;
+}
+.error {
+  background-color: #c51f1f;
+  margin-top: 8px;
+  padding-top: 6px;
+  padding-bottom: 6px;
+  border-radius: 15px;
+  color: white;
+  text-align: center;
+}
