+++ src/app/pages/gh-simulator/gh-simulator.component.css
@@ -1,43 +1,59 @@
-.token-section div {
-  margin: 20px;
-  padding: 20px;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-  background-color: #f9f9f9;
+.btn-back, .btn-edit {
+  background-color: white;
+  color: #007BFF;
+  border: 2px solid #007BFF;
+  margin: 10px;
 }
 
-.token-section p {
-  font-size: 18px;
-  color: #333;
+.btn-back {
+  margin-bottom: 15px!important;
 }
 
-.token-section button {
-  margin-top: 10px;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 5px;
-  background-color: #007bff;
-  color: #fff;
-  cursor: pointer;
-}
-
-.token-section button:hover {
-  background-color: #0056b3;
-}
-
-.token-section input {
-  margin-top: 10px;
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-  width: 100%;
-}
-.error {
-  background-color: #c51f1f;
-  margin-top: 8px;
-  padding-top: 6px;
-  padding-bottom: 6px;
-  border-radius: 15px;
+.btn-edit {
+  background-color: #6da7e7;
   color: white;
-  text-align: center;
+  justify-content: center;
+  position: absolute;
+  bottom: 0;
+  width: 50%;
+  font-size: 20px;
+  padding-top: 10px;
+  margin-bottom: 30px;
 }
+
+pre, textarea {
+  overflow: auto;
+  white-space: pre-wrap;
+  height: 500px;
+  margin-bottom: 10px;
+}
+
+pre {
+  height: 600px;
+}
+
+h4 {
+  background-color: lightblue;
+}
+
+.col-12, .card-full-height, .card-body {
+  height: 100%;
+  margin-bottom: 20px;
+}
+
+.card-full-height, .card-body {
+
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.card-full-height {
+  position: relative;
+  height: 280px;
+}
+
+.alert-warning {
+  margin-left: 20%;
+  margin-right: 20%;
+}
+
