+++ src/app/pages/gh-simulator/local/actions/actions.component.html
@@ -1,0 +1,1 @@
+<p>actions works!</p>
