+++ src/app/pages/gh-simulator/local/local.component.html
@@ -1,0 +1,1 @@
+<p>local works!</p>
