+++ src/app/pages/gh-simulator/local/local.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-local',
+  templateUrl: './local.component.html',
+  styleUrls: ['./local.component.css']
+})
+export class LocalComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
